certain fixes
