+++ Makefile
@@ -6,7 +6,7 @@
 #    By: tjmari <tjmari@student.42.fr>              +#+  +:+       +#+         #
 #                                                 +#+#+#+#+#+   +#+            #
 #    Created: 2021/05/22 20:07:24 by tjmari            #+#    #+#              #
-#    Updated: 2021/05/30 18:29:37 by tjmari           ###   ########.fr        #
+#    Updated: 2021/06/02 11:04:43 by tjmari           ###   ########.fr        #
 #                                                                              #
 # **************************************************************************** #
 
@@ -28,7 +28,7 @@ SRCS = ./srcs/main.c ./srcs/get_next_line.c \
 	./srcs/m_finalize.c  ./srcs/m_err_shrtn.c \
 	./srcs/m_fin_tools.c \
 	\
-	./srcs/t_executing.c ./srcs/t_signals.c \
+	./srcs/t_init.c ./srcs/t_executing.c ./srcs/t_signals.c \
 	./srcs/t_redirection.c ./srcs/t_pipe.c ./srcs/t_infork.c \
 	./srcs/t_builtin.c ./srcs/t_echo.c ./srcs/t_cd.c \
 	./srcs/t_pwd.c ./srcs/t_export.c ./srcs/t_unset.c \
