+++ srcs/t_charp.c
@@ -6,7 +6,7 @@
 /*   By: tjmari <tjmari@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/05/26 11:20:31 by tjmari            #+#    #+#             */
-/*   Updated: 2021/05/30 18:20:14 by tjmari           ###   ########.fr       */
+/*   Updated: 2021/06/02 11:04:12 by tjmari           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -92,36 +92,4 @@ char	**change_envp_var(char **argv, int index, char *arg)
 	free(argv[index]);
 	argv[index] = ft_strdup(arg);
 	return (argv);
-}
-
-void	change_shlvl(void)
-{
-	int		i;
-	char	**parts;
-	char	*temp;
-	int		shlvl;
-
-	i = 0;
-	while (i < how_many_element(g_tool.envp))
-	{
-		parts = ft_split(g_tool.envp[i], '=');
-		if (!(ft_strcmp(parts[0], "SHLVL")))
-		{
-			shlvl = ft_atoi(parts[1]);
-			shlvl++;
-			parts[1] = ft_itoa(shlvl);
-			temp = ft_strjoin("", parts[0]);
-			temp = ft_strjoin(temp, "=");
-			temp = ft_strjoin(temp, parts[1]);
-			g_tool.envp[i] = temp;
-			return ;
-		}
-		i++;
-		if (!g_tool.envp[i])
-		{
-			parts = ft_split("export SHLVL=1", ' ');
-			g_tool.envp = add_node_dc(g_tool.envp, parts, 2);
-			return ;
-		}
-	}
 }
