+++ srcs/t_env.c
@@ -6,7 +6,7 @@
 /*   By: tjmari <tjmari@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/05/25 20:12:47 by tjmari            #+#    #+#             */
-/*   Updated: 2021/06/01 18:48:28 by tjmari           ###   ########.fr       */
+/*   Updated: 2021/06/02 10:47:38 by tjmari           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -25,7 +25,10 @@ void	ft_env(void)
 		key = ft_getkey(g_tool.envp[i]);
 		value = ft_getvalue(g_tool.envp[i]);
 		if (!ft_strchr(g_tool.envp[i], '='))
+		{
+			i++;
 			continue ;
+		}
 		else if (ft_strcmp(ft_getvalue(g_tool.envp[i]), ""))
 			printf("%s=%s\n", key, value);
 		else
